Organize routes for info pages
