+++ src/router.js
@@ -4,9 +4,11 @@ import Vue from 'vue'
 import VueRouter from 'vue-router'
 
 import Home from './views/Home.vue'
-import About from './views/About.vue'
 import Profile from './views/Profile.vue'
 import AdminOnly from './views/AdminOnly.vue'
+
+import About from './views/info/About.vue'
+import ReservationInfo from './views/info/ReservationInfo.vue'
 
 // Observatories
 import Site from './views/Site.vue'
@@ -38,8 +40,12 @@ const router = new VueRouter({
       beforeEnter: authGuard, 
       meta: { requiresRole: 'admin'}
     },
-    { path: '/profile', name: 'profile', component: Profile, beforeEnter: authGuard,},
+
     { path: '/about', name: 'about', component: About },
+    { path: '/info/reservations', name: 'reservations', component: ReservationInfo},
+
+
+    { path: '/profile', name: 'profile', component: Profile, beforeEnter: authGuard,},
     { path: '/btns', name: 'btns', component: btns },
     { path: '/skymap', name: 'skymap', component: skymap },
     { path: '/analysis', name: 'analysis', component: analysis },
