+++ src/views/info/ReservationInfo.vue
@@ -1,0 +1,12 @@
+<template>
+  <div class="container main">
+    <h1 class="title">Reservations</h1>
+    <p>This is a page that will describe how reservations work.</p>
+  </div>
+</template>
+
+<style scoped>
+.main {
+  margin-top: 1em;
+}
+</style>
